refactor: 유저마이페이지, 나의쿠폰보기 수정
