+++ src/main/java/com/sparta/popupstore/domain/promotionevent/repository/CouponRepository.java
@@ -1,0 +1,10 @@
+package com.sparta.popupstore.domain.promotionevent.repository;
+
+import com.sparta.popupstore.domain.promotionevent.entity.Coupon;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
+
+    List<Coupon> findByUserId(Long userId);
+}
